+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.9.2] - 2019-08-01
+
+### Fixed
+- Fixed type definition to include `withApp` ([#125])
+
+
+## [0.9.1] - 2019-07-31
+
+### Fixed
+- Fixed renderer to be secondary when using `<Stage />` component
+
+
 ## [0.9.0] - 2019-07-01
 
 ### Added
@@ -258,7 +270,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `<TilingSprite />` component
 
 
-[Unreleased]: https://github.com/michalochman/react-pixi-fiber/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/michalochman/react-pixi-fiber/compare/v0.9.2...HEAD
+[0.9.2]: https://github.com/michalochman/react-pixi-fiber/compare/v0.9.1...v0.9.2
+[0.9.1]: https://github.com/michalochman/react-pixi-fiber/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/michalochman/react-pixi-fiber/compare/v0.8.2...v0.9.0
 [0.8.2]: https://github.com/michalochman/react-pixi-fiber/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/michalochman/react-pixi-fiber/compare/v0.8.0...v0.8.1
@@ -289,6 +303,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [0.2.0]: https://github.com/michalochman/react-pixi-fiber/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/michalochman/react-pixi-fiber/compare/v0.1.0...v0.1.1
 
+[#125]: https://github.com/michalochman/react-pixi-fiber/issues/125
 [#122]: https://github.com/michalochman/react-pixi-fiber/issues/122
 [#120]: https://github.com/michalochman/react-pixi-fiber/issues/120
 [#118]: https://github.com/michalochman/react-pixi-fiber/issues/118